Public-release prep: docs, MIT license, and security hardening (#76)
* docs: add license, contribution guide, and prep repo for public release

Add MIT LICENSE, CONTRIBUTING.md (issue-first workflow), SECURITY.md
(private reporting + the intentional no-auth model), and GitHub issue/PR
templates. Rewrite README to lead with the published, signed GHCR image
and add Verifying/Updating/Security-model/Backup sections. Document the
movies feature (README + CLAUDE.md) and the DB_MAX_CONNECTIONS env var.

Remove an accidentally-committed SQLite test artifact
("backend/not a sqlite url with space", schema-only, zero rows) and
gitignore the Playwright scratch dir.


* fix(security): prevent secret leakage, harden supply chain and headers

Close a high-severity leak where the TMDB API key and Slack webhook URL
could reach unauthenticated /sync and /notifications/test response bodies
(and logs) via reqwest error strings that embed the request URL. Strip the
URL centrally in From<reqwest::Error> via without_url(), and add
AppError::client_message() so internal error variants never serialize into
a response body. (#70)

Supply chain: pin every GitHub Action to a commit SHA and add
helpers:pinGitHubActionDigests (#71); build the frontend image from the
audited lockfile on node:24-slim with `npm ci --ignore-scripts` and add
`cargo build --locked` (#72); scope ci.yml permissions, pin the Socket CLI,
narrow `secrets: inherit`, and give Renovate digest auto-merge a soak (#73).

Hardening: add CSP + X-Content-Type-Options/X-Frame-Options/Referrer-Policy
response headers (#74); reject NOTIFICATION_CHECK_INTERVAL_MINUTES=0 instead
of panicking, widen notification season/episode to i64, cap TMDB response
size, and quote/encode TMDB-derived URLs in the frontend (#75).


---------

+++ backend/src/notifications/mod.rs
@@ -10,8 +10,8 @@ use crate::error::Result;
 pub enum NotificationEvent {
     EpisodeAiringToday {
         show_name: String,
-        season: u16,
-        episode: u16,
+        season: i64,
+        episode: i64,
         title: String,
         where_to_watch: Vec<String>,
     },
